requirements.txt
# Default
rank-bm25
python-dotenv
sentence_transformers
streamlit
llama-index
qdrant_client
llama-index-embeddings-huggingface
llama_index.embeddings.langchain
fastembed
llama-index-llms-groq